Backend/requirements.txt
asgiref==3.5.2
certifi==2022.6.15.1
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
defusedxml==0.7.1
Django==4.1.1
django-allauth==0.51.0
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
djongo==1.3.6
dnspython==2.2.1
haversine==2.7.0
idna==3.3
joblib==1.1.0
numpy==1.22.2
oauthlib==3.2.0
pandas==1.4.4
pycparser==2.21
PyJWT==2.4.0
pymongo==3.12.3
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
scikit-learn==1.1.2
scipy==1.9.1
six==1.16.0
sqlparse==0.2.4
threadpoolctl==3.1.0
tzdata==2022.2
urllib3==1.26.12
